requirements.txt
accelerate==0.33.0
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
av==12.3.0
certifi==2024.7.4
charset-normalizer==3.3.2
datasets==2.20.0
dill==0.3.8
evaluate==0.4.2
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.5.0
fvcore==0.1.5.post20221221
huggingface-hub==0.24.0
idna==3.7
iopath==0.1.10
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numpy==1.26.4
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
pillow==10.4.0
portalocker==2.10.1
psutil==6.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
python-dateutil==2.9.0.post0
pytorchvideo==0.1.5
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
six==1.16.0
sympy==1.13.1
tabulate==0.9.0
termcolor==2.4.0
tokenizers==0.19.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
transformers==4.42.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.4
